settingPage
구조 정리 scss 입히기
