+++ web/src/component/Setting/index.module.scss
@@ -1,0 +1,133 @@
+.index:global {
+    > .settingContainer{
+        > .settingTitle {
+            font-family: "Inter";
+            font-style: normal;
+            font-weight: 700;
+            font-size: 34px;
+            line-height: 41px;
+            margin: 15px 15px 15px 15px;
+        }
+        > .settingBoxContainer {
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+
+            > .settingBox {
+                width: 345px;
+                height: 50px;
+                background-color: #eaeaea;
+                border-radius: 8px;
+                margin-bottom: 15px;
+                
+                >   
+                .settingBoxPadding {
+                    display: flex;
+                    /* flex-direction: row; */
+                    justify-content: space-between;
+                    align-items: center;
+                    /* padding: 13px 16px 13px 16px; */
+                    /* padding-top: 13px; */
+                    /* padding-bottom: 13px; */
+
+                    > .settingText {
+                        font-family: "Inter";
+                        font-style: normal;
+                        font-weight: 700;
+                        font-size: 16px;
+                        line-height: 24px;
+                        padding: 13px 0px 13px 16px;
+                    }
+                    > .rightArrowIcon {
+                        width: 11px;
+                        height: 22px;
+                        padding: 13px 16px 13px 0px;
+                    }
+                    > .toggle-container-{
+                        background-color: #c4c4c4;
+                        cursor: pointer;
+                        user-select: none;
+                        border-radius: 25px;
+                        margin: 9px 8px 10px 0px;
+                        /* padding: 2px; */
+                        height: 31px;
+                        width: 51px;
+                        position: relative;
+                        
+                        > .dialog-button-disabled{
+                            font-size: 14px;
+                            line-height: 16px;
+                            cursor: pointer;
+                            background-color: #ffffff;
+                            color: white;
+                            padding: 8px 12px;
+                            /* border-radius: 25px; */
+                            /* min-width: 46px; */
+                            /* display: flex;
+                            justify-content: center;
+                            align-items: center; */
+                            height: 27px;
+                            width: 27px;
+                            min-width: unset;
+                            border-radius: 25px;
+                            box-sizing: border-box;
+                            box-shadow: 0px 3px 8px rgba(0, 0, 0, 0.15), 0px 3px 1px rgba(0, 0, 0, 0.06);
+                            position: absolute;
+                            left: 1px;
+                            top: 2px;
+                            transition: all 0.3s ease;
+                        }
+                    }
+                    > .toggle-container-enabled {
+                        background-color: #34c759;
+                        cursor: pointer;
+                        user-select: none;
+                        border-radius: 25px;
+                        margin: 9px 8px 10px 0px;
+                        /* padding: 2px; */
+                        height: 31px;
+                        width: 51px;
+                        position: relative;
+
+                        > .dialog-button- {
+                            font-size: 14px;
+                            line-height: 16px;
+                            cursor: pointer;
+                            background-color: #ffffff;
+                            color: white;
+                            padding: 8px 12px;
+                            /* border-radius: 25px; */
+                            /* min-width: 46px; */
+                            /* display: flex;
+                            justify-content: center;
+                            align-items: center; */
+                            height: 27px;
+                            width: 27px;
+                            min-width: unset;
+                            border-radius: 25px;
+                            box-sizing: border-box;
+                            box-shadow: 0px 3px 8px rgba(0, 0, 0, 0.15), 0px 3px 1px rgba(0, 0, 0, 0.06);
+                            position: absolute;
+                            left: 23px;
+                            top: 2px;
+                            transition: all 0.3s ease;
+                        }
+
+                    }
+                }
+            }
+        }
+    }
+    
+ 
+  
+
+
+
+  
+
+  
+
+
+}
